feat(ui): add LegacyApp for graph and bug reproduction workflows
Introduces a `LegacyApp` frontend component with support for knowledge graph visualization, PR analysis, and bug reproduction workflows. Includes comprehensive state management, error handling, and integration of various UI panels for navigation, analysis, and bug reporting.
