+++ backend/tests/conftest.py
@@ -3,7 +3,6 @@ import json
 import httpx
 import pytest
 
-import indexer
 from agent.bug_run_registry import clear_bug_runs
 from agent.sessions import clear_sessions
 from main import app
@@ -19,12 +18,10 @@ async def client():
 
 
 @pytest.fixture(autouse=True)
-def reset_registry():
-    indexer._registry.clear()
+def reset_sessions():
     clear_sessions()
     clear_bug_runs()
     yield
-    indexer._registry.clear()
     clear_sessions()
     clear_bug_runs()
