+++ backend/tests/indexer/test_indexer_async.py
@@ -1,7 +1,12 @@
 from unittest.mock import AsyncMock, MagicMock, patch
+from uuid import uuid4
 
-import indexer
-from indexer import get_graph_data, index_repo
+import pytest
+
+import db
+from tests.fake_supabase import FakeSupabaseClient
+
+from indexer import get_graph_data, index_repo, list_indexed_repos
 from models import (
     ErrorEvent,
     GraphCluster,
@@ -11,6 +16,17 @@ from models import (
     IndexStepEvent,
 )
 
+
+# ── fixtures ──────────────────────────────────────────────────────────────────
+
+@pytest.fixture
+def fake_supabase(monkeypatch):
+    fake = FakeSupabaseClient()
+    monkeypatch.setattr(db, "get_client", lambda: fake)
+    return fake
+
+
+# ── helpers ───────────────────────────────────────────────────────────────────
 
 def make_git_proc(returncode=0, stderr=b""):
     proc = MagicMock()
@@ -38,28 +54,31 @@ async def collect(gen):
 
 # --- index_repo ---
 
-async def test_index_repo_bad_url_yields_error():
-    events = await collect(index_repo("https://github.com/onlyone"))
+async def test_index_repo_bad_url_yields_error(fake_supabase):
+    uid = uuid4()
+    events = await collect(index_repo(uid, "https://github.com/onlyone"))
     assert len(events) == 1
     assert isinstance(events[0], ErrorEvent)
     assert "Cannot parse" in events[0].message
 
 
-async def test_index_repo_git_clone_failure():
+async def test_index_repo_git_clone_failure(fake_supabase):
+    uid = uuid4()
     git_proc = make_git_proc(returncode=1, stderr=b"repository not found")
 
     async def mock_exec(*args, **kwargs):
         return git_proc
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
-        events = await collect(index_repo("https://github.com/owner/repo"))
+        events = await collect(index_repo(uid, "https://github.com/owner/repo"))
 
     error_events = [e for e in events if isinstance(e, ErrorEvent)]
     assert len(error_events) == 1
     assert "git clone failed" in error_events[0].message
 
 
-async def test_index_repo_gitnexus_not_found():
+async def test_index_repo_gitnexus_not_found(fake_supabase):
+    uid = uuid4()
     git_proc = make_git_proc(returncode=0)
 
     async def mock_exec(*args, **kwargs):
@@ -68,14 +87,15 @@ async def test_index_repo_gitnexus_not_found():
         raise FileNotFoundError("gitnexus not found")
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
-        events = await collect(index_repo("https://github.com/owner/repo"))
+        events = await collect(index_repo(uid, "https://github.com/owner/repo"))
 
     error_events = [e for e in events if isinstance(e, ErrorEvent)]
     assert len(error_events) == 1
     assert "gitnexus not found" in error_events[0].message
 
 
-async def test_index_repo_successful_flow_events():
+async def test_index_repo_successful_flow_events(fake_supabase):
+    uid = uuid4()
     git_proc = make_git_proc(returncode=0)
     gitnexus_proc = make_gitnexus_proc(
         lines=["Parsing files\n", "Clustering nodes\n"], returncode=0
@@ -89,7 +109,7 @@ async def test_index_repo_successful_flow_events():
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
         with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
-            events = await collect(index_repo("https://github.com/owner/myrepo"))
+            events = await collect(index_repo(uid, "https://github.com/owner/myrepo"))
 
     assert any(isinstance(e, IndexStepEvent) for e in events)
     assert any(isinstance(e, IndexDoneEvent) for e in events)
@@ -99,7 +119,8 @@ async def test_index_repo_successful_flow_events():
     assert done.repo == "owner/myrepo"
 
 
-async def test_index_repo_successful_flow_event_order():
+async def test_index_repo_successful_flow_event_order(fake_supabase):
+    uid = uuid4()
     git_proc = make_git_proc(returncode=0)
     gitnexus_proc = make_gitnexus_proc(lines=[], returncode=0)
     mock_graph = GraphData(nodes=[], edges=[], clusters=[])
@@ -111,13 +132,15 @@ async def test_index_repo_successful_flow_event_order():
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
         with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
-            events = await collect(index_repo("https://github.com/owner/repo"))
+            events = await collect(index_repo(uid, "https://github.com/owner/repo"))
 
     # IndexDoneEvent must come last
     assert isinstance(events[-1], IndexDoneEvent)
 
 
-async def test_index_repo_populates_registry():
+async def test_index_repo_creates_project_record(fake_supabase):
+    """After successful indexing, the project should exist in the DB with status 'ready'."""
+    uid = uuid4()
     git_proc = make_git_proc(returncode=0)
     gitnexus_proc = make_gitnexus_proc(lines=[], returncode=0)
     mock_graph = GraphData(nodes=[], edges=[], clusters=[])
@@ -129,53 +152,50 @@ async def test_index_repo_populates_registry():
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
         with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
-            await collect(index_repo("https://github.com/owner/myrepo"))
+            await collect(index_repo(uid, "https://github.com/owner/myrepo"))
 
-    assert "owner/myrepo" in indexer._registry
+    repos = list_indexed_repos(uid)
+    assert len(repos) == 1
+    assert repos[0]["owner"] == "owner"
+    assert repos[0]["repo_name"] == "myrepo"
+    assert repos[0]["index_status"] == "ready"
 
 
 # --- get_graph_data ---
 
-async def test_get_graph_data_not_indexed():
-    result = await get_graph_data("nobody", "norepo")
+async def test_get_graph_data_not_indexed(fake_supabase):
+    uid = uuid4()
+    result = await get_graph_data(uid, "nobody", "norepo")
     assert result.nodes == []
     assert result.edges == []
     assert result.clusters == []
 
 
-async def test_get_graph_data_cached_graph():
-    cached = GraphData(
+async def test_get_graph_data_returns_live_graph(fake_supabase):
+    """get_graph_data fetches live from MCP for an existing project."""
+    uid = uuid4()
+    # Pre-create a project record so get_graph_data finds it
+    from projects import create_project
+    create_project(uid, "https://github.com/owner/repo")
+
+    mock_graph = GraphData(
         nodes=[GraphNode(id="n1", label="f.py", type="file", cluster="c")],
         edges=[],
         clusters=[GraphCluster(id="c", label="C", size=1)],
     )
-    indexer._registry["owner/repo"] = {"path": "/fake", "repo_name": "repo", "graph": cached}
 
-    result = await get_graph_data("owner", "repo")
+    with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
+        result = await get_graph_data(uid, "owner", "repo")
+
     assert len(result.nodes) == 1
     assert result.nodes[0].id == "n1"
 
 
-async def test_get_graph_data_lazy_read():
-    mock_graph = GraphData(
-        nodes=[GraphNode(id="n2", label="g.py", type="file", cluster="c")],
-        edges=[],
-        clusters=[],
-    )
-    indexer._registry["owner/repo2"] = {"path": "/fake", "repo_name": "repo2", "graph": None}
-
-    with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
-        result = await get_graph_data("owner", "repo2")
-
-    assert len(result.nodes) == 1
-    # Result is now cached in registry
-    assert indexer._registry["owner/repo2"]["graph"] is not None
-
-
 # --- embeddings ---
 
-async def test_embeddings_flag_passed():
+async def test_embeddings_flag_passed(fake_supabase):
     """gitnexus analyze must be called with --embeddings."""
+    uid = uuid4()
     git_proc = make_git_proc(returncode=0)
     gitnexus_proc = make_gitnexus_proc(lines=[], returncode=0)
     mock_graph = GraphData(nodes=[], edges=[], clusters=[])
@@ -189,7 +209,7 @@ async def test_embeddings_flag_passed():
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
         with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
-            await collect(index_repo("https://github.com/owner/repo"))
+            await collect(index_repo(uid, "https://github.com/owner/repo"))
 
     gitnexus_calls = [a for a in captured_args if a[0] == "gitnexus"]
     assert gitnexus_calls, "gitnexus was never called"
@@ -198,8 +218,9 @@ async def test_embeddings_flag_passed():
     )
 
 
-async def test_embeddings_stage_emitted():
+async def test_embeddings_stage_emitted(fake_supabase):
     """A stdout line containing 'embedding' must yield IndexStepEvent(stage='embeddings')."""
+    uid = uuid4()
     git_proc = make_git_proc(returncode=0)
     gitnexus_proc = make_gitnexus_proc(
         lines=["Generating embeddings for 42 files\n"], returncode=0
@@ -213,7 +234,7 @@ async def test_embeddings_stage_emitted():
 
     with patch("asyncio.create_subprocess_exec", mock_exec):
         with patch("indexer._fetch_stats_and_graph", return_value=({}, mock_graph)):
-            events = await collect(index_repo("https://github.com/owner/repo"))
+            events = await collect(index_repo(uid, "https://github.com/owner/repo"))
 
     embedding_events = [
         e for e in events
